refactor: split profiles/commands.ts into focused modules
commands.ts (901 lines) split into:
- profiles/shared.ts      — constants, validation, dir helpers, manifest fetch/verify
- profiles/cmd-local.ts   — list, remove, feed, check (no network required)
- profiles/cmd-catalog.ts — install, update, seed, info, available (manifest network ops)
- profiles/cmd-test.ts    — testProfile, cmdTest (diagnostic subcommand)
- profiles/commands.ts    — dispatcher + re-exports (67 lines)

All external imports unchanged; tests import from commands.ts as before.
